Refactor wallpaper functionality
